pull

+++ app/src/main/java/com/example/dzj/android_practice/kotlindemo/CommonUtil.kt
@@ -2,7 +2,7 @@ package com.example.dzj.android_practice.kotlindemo
 
 import android.content.res.Resources
 
-fun dpToPx(dp:Int): Float = Resources.getSystem().displayMetrics.density * dp
+fun dpToPx(dp:Float): Float = Resources.getSystem().displayMetrics.density * dp
 
 val <T> List<T>.lastData:T
 get() {
